fix(app): fixed style flash in about page
